+++ Runtime/ContinuousDistributions/DistributionFilters/FilteredGenerators/ContinuousFilteredGenerator.cs
@@ -61,7 +61,7 @@ namespace Zor.RandomGenerators.ContinuousDistributions.DistributionFilters
 			{
 				m_sequence[m_currentSequenceLength++] = generated;
 			}
-			else
+			else if (m_currentSequenceLength > 0)
 			{
 				int last = m_currentSequenceLength - 1;
 				Array.Copy(m_sequence, 1, m_sequence, 0, last);