+++ Runtime/ContinuousDistributions/DistributionFilters/GreaterFilters/FilterProviders/GreaterFilterProvider.cs
@@ -6,40 +6,40 @@ using UnityEngine;
 namespace Zor.RandomGenerators.ContinuousDistributions.DistributionFilters
 {
 	[CreateAssetMenu(
-		menuName = CreateAssetMenuConstants.GreaterSequenceFiltersFolder + "Greater Sequence Filter Provider",
+		menuName = CreateAssetMenuConstants.GreaterContinuousFiltersFolder + "Greater Sequence Filter Provider",
 		fileName = "Greater Sequence Filter Provider",
 		order = CreateAssetMenuConstants.FilterOrder
 	)]
-	public sealed class GreaterSequenceFilterProvider : ContinuousFilterProvider
+	public sealed class GreaterFilterProvider : ContinuousFilterProvider
 	{
 #pragma warning disable CS0649
-		[SerializeField] private GreaterSequenceFilter m_GreaterSequenceFilter;
+		[SerializeField] private GreaterFilter m_GreaterFilter;
 #pragma warning restore CS0649
 
 		public override IContinuousFilter filter
 		{
 			[Pure]
-			get => new GreaterSequenceFilter(m_GreaterSequenceFilter);
+			get => new GreaterFilter(m_GreaterFilter);
 		}
 
 		public override IContinuousFilter sharedFilter
 		{
 			[Pure]
-			get => m_GreaterSequenceFilter;
+			get => m_GreaterFilter;
 		}
 
 		[NotNull]
-		public GreaterSequenceFilter greaterSequenceFilter
+		public GreaterFilter greaterFilter
 		{
 			[Pure]
-			get => new GreaterSequenceFilter(m_GreaterSequenceFilter);
+			get => new GreaterFilter(m_GreaterFilter);
 		}
 
 		[NotNull]
-		public GreaterSequenceFilter sharedGreaterSequenceFilter
+		public GreaterFilter sharedGreaterFilter
 		{
 			[Pure]
-			get => m_GreaterSequenceFilter;
+			get => m_GreaterFilter;
 		}
 	}
 }